Load current user from the main component
This allow to have the current user in any page, especially in 404 page.
Change Main component into container.
