--- src/jsx/components/Main.js
@@ -1,21 +1,0 @@
-import React, { Component } from 'react'
-import HeaderPage from '../containers/HeaderPage' 
-import Footer from './Footer' 
-import PlayerBoxPage from '../containers/PlayerBoxPage'
-
-class Main extends Component {
-    render() {
-        return (
-            <div id="main">
-                <HeaderPage/>
-                <div id="content">
-                    <PlayerBoxPage/>
-                    {this.props.children}
-                </div>
-                <Footer/>
-            </div>
-        )
-    }
-}
-
-export default Main
