+++ src/jsx/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { Router, Route, Redirect, IndexRedirect, browserHistory } from 'react-router'
-import Main from './components/Main'
+import Main from './containers/Main'
 import LoggedinPage from './containers/LoggedinPage'
 import LoginForm from './containers/LoginForm'
 import LogoutPage from './containers/LogoutPage'
@@ -13,6 +13,7 @@ import LibraryListSongPage from './containers/LibraryListSongPage'
 import LibraryListArtistPage from './containers/LibraryListArtistPage'
 import LibraryListWorkPage from './containers/LibraryListWorkPage'
 import NotFound from './components/NotFound'
+import Forbidden from './components/Forbidden'
 import NotFoundRedirector from './components/NotFoundRedirector'
 import { createStore, applyMiddleware, compose } from 'redux'
 import { Provider } from 'react-redux'
@@ -50,6 +51,8 @@ ReactDOM.render(
                     <Route path="user" component={UserPage}/>
                     <Route path="users" component={UsersPage}/>
                     <Route path="users/:userId" component={UserEditPage}/>
+                    <Route path="403" component={Forbidden}/>
+                    {/* <Route path="404" component={NotFound}/> */}
                 </Route>
                 <Route path="login" component={LoginForm}/>
                 <Route path="logout" component={LogoutPage}/>
